Centralise SHA512 (#19)
* centralises sha512

* updates sntrup4591761 benchmarks
