--- crypto_kem/sntrup4591761/ref/crypto_hashblocks_sha512.h
@@ -1,8 +1,0 @@
-
-#ifndef crypto_hashblocks_sha512_h
-#define crypto_hashblocks_sha512_h
-
-int crypto_hashblocks_sha512(unsigned char *out,const unsigned char *in,unsigned long long inlen);
-#endif
-
-
